Clib: Correct memset() declarations.
Gcc complains memset() prototype used in the ACPICA even when the nostdinc
is specified:
  ../../../source/include/acutils.h:418:1: error: conflicting types for built-in function 'memset' [-Werror]

This patch corrects the declarations of memset(). Lv Zheng.

+++ source/components/utilities/utclib.c
@@ -257,7 +257,7 @@ memcpy (
 void *
 memset (
     void                    *Dest,
-    UINT8                   Value,
+    int                     Value,
     ACPI_SIZE               Count)
 {
     char                    *New = (char *) Dest;
